Refactor fusion reaction handling to use dedicated module
